refactor: move AgendaDetailsScreenState to separate file
